==> analysis/06a-DAanalysis.Rmd <==
---
title: "Cell type composition analysis"
author: "Belinda Phipson"
date: "10/02/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    message = FALSE,
    warning = FALSE
)
```

# Introduction

Here we test whether the cell type composition of the heart, young and adult 
samples differ using the `propeller` function in the `speckle` package.

# Load libraries and functions

```{r}
library(edgeR)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(Seurat)
library(monocle)
library(cowplot)
library(DelayedArray)
library(scran)
library(NMF)
library(workflowr)
library(ggplot2)
library(clustree)
library(dplyr)
library(speckle)
```

# Load the heart data

```{r}
heart <- readRDS(file="./output/heart-int-FYA.Rds")
```

# Set up cell information for propeller analysis
```{r}
heart$sample <- factor(heart$biorep, levels=c(paste("f",1:3, sep=""),
                                              paste("y",1:3, sep=""),
                                              paste("a",1:3, sep="")))
heart$group <- NA 
heart$group[grep("f",heart$sample)] <- "fetal"
heart$group[grep("y",heart$sample)] <- "young"
heart$group[grep("a",heart$sample)] <- "adult"
heart$group <- factor(heart$group, levels=c("fetal", "young","adult"))
```

# Visualise the data 

## tSNE plots

```{r}
DimPlot(heart, reduction="tsne", group.by = "Broad_celltype")
```

```{r, fig.width=12,fig.height=6}
DimPlot(heart, reduction="tsne", group.by = "Broad_celltype", split.by="group")
```

## Barplots of proportions

```{r}
plotCellTypeProps(clusters=heart$Broad_celltype, sample=heart$sample)
```

## Biological variability plots

These plots show that there is a massive amount of biological variability 
between the samples, hence using Poisson or binomial models are not appropriate.

```{r, fig.width=13,fig.height=6}
# get the cell type counts and proportions
x <- getTransformedProps(clusters = heart$Broad_celltype, sample=heart$sample,
                         transform="logit")
par(mfrow=c(1,2))
plotCellTypeMeanVar(x$Counts)
plotCellTypePropsMeanVar(x$Counts)
```

# Testing for differences in proportions

```{r}
Idents(heart) <- heart$Broad_celltype
```

```{r}
out <- propeller(heart, transform = "logit")
out
# Significant cell types at FDR 0.05
rownames(out)[which(out$FDR<0.05)]
```

# Visualise the results

```{r}
# Set up group information based on counts matrix/matrix of proportions
group <- factor(rep(c("fetal","young","adult"), each=3), 
                levels=c("fetal","young","adult"))
```


```{r, fig.width=10, fig.height=10}
ct <- rownames(out)
par(mfrow=c(3,3))
for(i in 1:nrow(out)){
  stripchart(x$Proportions[ct[i],]~group, vertical=TRUE, pch=16, 
             method="jitter", ylab="Proportion", main=ct[i], 
             col=ggplotColors(3), cex=1.5, cex.lab=1.5, cex.axis=1.5,
             cex.main=1.5)
}

```
