==> analysis/11-ClustEndo.Rmd <==
---
title: "Recluster endothelial cells"
author: "Belinda Phipson"
date: "10/31/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    message = FALSE,
    warning = FALSE
)
```

# Load libraries and functions

```{r}
library(edgeR)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(Seurat)
library(monocle)
library(cowplot)
library(DelayedArray)
library(scran)
library(NMF)
library(workflowr)
library(ggplot2)
library(clustree)
library(dplyr)
```

```{r echo=FALSE}
source("/Users/phipsonbelinda/Documents/MCRI_collab/Porrello/Human_Development_snRNAseq/code/normCounts.R")
source("/Users/phipsonbelinda/Documents/MCRI_collab/Porrello/Human_Development_snRNAseq/code/findModes.R")
source("/Users/phipsonbelinda/Documents/MCRI_collab/Porrello/Human_Development_snRNAseq/code/ggplotColors.R")
```

```{r readTargets}
targets <- read.delim("/Users/phipsonbelinda/Documents/MCRI_collab/Porrello/Human_Development_snRNAseq/data/targets.txt",header=TRUE, stringsAsFactors = FALSE)
targets$FileName2 <- paste(targets$FileName,"/",sep="")
targets$Group_ID2 <- gsub("LV_","",targets$Group_ID)
group <- c("Fetal_1","Fetal_2","Fetal_3",
           "Young_1","Young_2","Young_3",
           "Adult_1","Adult_2","Adult_3", 
           "Diseased_1","Diseased_2",
           "Diseased_3","Diseased_4")
m <- match(group, targets$Group_ID2)
targets <- targets[m,]
```

```{r eval=FALSE}
fetal.integrated <- readRDS(file="./output/fetal-int.Rds")
load(file="./output/fetalObjs.Rdata")

young.integrated <- readRDS(file="./output/young-int.Rds")
load(file="./output/youngObjs.Rdata")

adult.integrated <- readRDS(file="./output/adult-int.Rds")
load(file="./output/adultObjs.Rdata")
```

# Set default clustering resolution

```{r eval=FALSE}
# Default 0.3
Idents(fetal.integrated) <- fetal.integrated$integrated_snn_res.0.3
DimPlot(fetal.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()

# Default 0.3
DimPlot(young.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()

# Default 0.6
DimPlot(adult.integrated, reduction = "tsne",label=TRUE,label.size = 6)+NoLegend()

```

# Merge all data together

```{r eval=FALSE}
heart <- merge(fetal.integrated, y = c(young.integrated, adult.integrated), project = "heart")
table(heart$orig.ident)
DefaultAssay(object = heart) <- "RNA"
```

# Get endothelial cells only

```{r eval=FALSE}
endo <- subset(heart,subset = Broad_celltype == "Endothelial cells")
dim(endo)
```

```{r echo=FALSE}
endo <- readRDS(file="./output/endo-int-FYA-filtered.Rds")
```

# Check for poor quality cells

Check for cells with very low number of uniquely detected genes. 

```{r}
par(mfrow=c(1,2))
plot(density(endo$nFeature_RNA),main="Number of genes detected")
abline(v=500,col=2)
plot(density(endo$nCount_RNA),main="Library size")
abline(v=2500,col=2)

#endo <- subset(endo, subset = nFeature_RNA > 500 & nCount_RNA > 2500)
dim(endo)
table(endo$biorep)
```


# Run new integration with SCtransform normalisation


```{r eval=FALSE}
endo.list <- SplitObject(endo, split.by = "biorep")
```

```{r eval=FALSE}
for (i in 1:length(endo.list)) {
    endo.list[[i]] <- SCTransform(endo.list[[i]], verbose = FALSE)
}
```

```{r eval=FALSE}
kf <- min(sapply(endo.list, ncol))
endo.anchors <- FindIntegrationAnchors(object.list = endo.list, dims=1:30,anchor.features = 3000,k.filter=kf)
endo.integrated <- IntegrateData(anchorset = endo.anchors,dims=1:30)
```

```{r echo=FALSE}
endo.integrated <- endo
```


# Perform clustering

```{r eval=TRUE}
DefaultAssay(object = endo.integrated) <- "integrated"
```

## Perform scaling and PCA

```{r eval=FALSE}
endo.integrated <- ScaleData(endo.integrated, verbose = FALSE)
endo.integrated <- RunPCA(endo.integrated, npcs = 50, verbose = FALSE)
```

```{r}
ElbowPlot(endo.integrated,ndims=50)
```

```{r,fig.height=10, fig.width=12}
VizDimLoadings(endo.integrated, dims = 1:4, reduction = "pca")
```

```{r}
DimPlot(endo.integrated, reduction = "pca",group.by="orig.ident")
DimPlot(endo.integrated, reduction = "pca",group.by="biorep")
DimPlot(endo.integrated, reduction = "pca",group.by="sex")
DimPlot(endo.integrated, reduction = "pca",group.by="batch")
```

```{r,fig.height=10, fig.width=12}
DimHeatmap(endo.integrated, dims = 1:15, cells = 500, balanced = TRUE)
DimHeatmap(endo.integrated, dims = 16:30, cells = 500, balanced = TRUE)
#DimHeatmap(endo.integrated, dims = 31:45, cells = 500, balanced = TRUE)
```

## Perform nearest neighbours clustering

```{r eval=FALSE}
endo.integrated <- FindNeighbors(endo.integrated, dims = 1:20)
endo.integrated <- FindClusters(endo.integrated, resolution = 0.1)
```

```{r}
table(Idents(endo.integrated))
```

```{r}
par(mfrow=c(1,1))
par(mar=c(5,4,2,2))
barplot(table(Idents(endo.integrated)),ylab="Number of cells",xlab="Clusters")
title("Number of cells in each cluster")
```


## Visualisation with TSNE

```{r eval=FALSE}
set.seed(10)
endo.integrated <- RunTSNE(endo.integrated, reduction = "pca", dims = 1:20)
```


```{r}
DimPlot(endo.integrated, reduction = "tsne",label=TRUE,label.size = 6,pt.size = 0.5)+NoLegend()
```

```{r eval=FALSE}
pdf(file="./output/Figures/tsne-endoALL-res01.pdf",width=10,height=8,onefile = FALSE)
DimPlot(endo.integrated, reduction = "tsne",label=TRUE,label.size = 6,pt.size = 0.5)+NoLegend()
dev.off()
```

```{r}
DimPlot(endo.integrated, reduction = "tsne", group.by = "orig.ident")
```

```{r,fig.width=12,fig.height=6}
DimPlot(endo.integrated, reduction = "tsne", split.by = "orig.ident")
```

```{r}
DimPlot(endo.integrated, reduction = "tsne", group.by = "biorep")
```


```{r}
DimPlot(endo.integrated, reduction = "tsne", group.by = "sex")
```

```{r,fig.width=10,fig.height=7}
DimPlot(endo.integrated, reduction = "tsne", split.by = "sex")
```

```{r}
DimPlot(endo.integrated, reduction = "tsne", group.by = "batch")
```


```{r,fig.width=12,fig.height=6}
par(mfrow=c(1,1))
par(mar=c(4,4,2,2))
tab <- table(Idents(endo.integrated),endo.integrated$biorep)
barplot(t(tab/rowSums(tab)),beside=TRUE,col=ggplotColors(9),legend=TRUE)
```

```{r,fig.width=12,fig.height=6}
par(mfrow=c(1,1))
par(mar=c(4,4,2,2))
tab <- table(Idents(endo.integrated),endo.integrated$orig.ident)
barplot(t(tab/rowSums(tab)),beside=TRUE,col=ggplotColors(3))
legend("topleft",legend=colnames(tab),fill=ggplotColors(3))
```

# Visualisation with clustree

```{r}
clusres <- c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2)
```
```{r eval=FALSE}
for(i in 1:length(clusres)){
  endo.integrated <- FindClusters(endo.integrated, 
                                   resolution = clusres[i])
}
```

```{r}
pct.male <- function(x) {mean(x=="m")}
pct.female <- function(x) {mean(x=="f")}
pct.fetal <- function(x) {mean(x=="fetal")}
pct.young <- function(x) {mean(x=="young")}
pct.adult <- function(x) {mean(x=="adult")}
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.")
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.",
         node_colour = "sex", node_colour_aggr = "pct.female",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.",
         node_colour = "sex", node_colour_aggr = "pct.male",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.fetal",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.young",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(endo.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.adult",assay="RNA")
```


# Save Seurat object

```{r}
DefaultAssay(endo.integrated) <- "RNA"
Idents(endo.integrated) <- endo.integrated$integrated_snn_res.0.1
```

```{r}
# saveRDS(endo.integrated,file="./output/RDataObjects/endo-int-FYA-filtered.Rds")
#endo.integrated <- readRDS(file="./output/RDataObjects/endo-int-FYA.Rds")
```

```{r}
# Load unfiltered counts matrix for every sample (object all)
#load("./output/RDataObjects/all-counts.Rdata")
```
