==> analysis/13-ClustNeuron.Rmd <==
---
title: "Recluster neuronal cells"
author: "Belinda Phipson"
date: "10/31/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    message = FALSE,
    warning = FALSE
)
```

# Load libraries and functions

```{r}
library(edgeR)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(Seurat)
library(monocle)
library(cowplot)
library(DelayedArray)
library(scran)
library(NMF)
library(workflowr)
library(ggplot2)
library(clustree)
library(dplyr)
```

```{r echo=FALSE}
source("./code/normCounts.R")
source("./code/findModes.R")
source("./code/ggplotColors.R")
```

```{r readTargets}
targets <- read.delim("./data/targets.txt",header=TRUE, stringsAsFactors = FALSE)
targets$FileName2 <- paste(targets$FileName,"/",sep="")
targets$Group_ID2 <- gsub("LV_","",targets$Group_ID)
group <- c("Fetal_1","Fetal_2","Fetal_3",
           "Young_1","Young_2","Young_3",
           "Adult_1","Adult_2","Adult_3", 
           "Diseased_1","Diseased_2",
           "Diseased_3","Diseased_4")
m <- match(group, targets$Group_ID2)
targets <- targets[m,]
```

```{r eval=FALSE}
fetal.integrated <- readRDS(file="./output/RDataObjects/fetal-int.Rds")
load(file="./output/RDataObjects/fetalObjs.Rdata")

young.integrated <- readRDS(file="./output/RDataObjects/young-int.Rds")
load(file="./output/RDataObjects/youngObjs.Rdata")

adult.integrated <- readRDS(file="./output/RDataObjects/adult-int.Rds")
load(file="./output/RDataObjects/adultObjs.Rdata")
```

# Merge all data together

```{r eval=FALSE}
heart <- merge(fetal.integrated, y = c(young.integrated, adult.integrated), project = "heart")
DefaultAssay(object = heart) <- "RNA"
```

# Get neuronal cells only

```{r eval=FALSE}
neuron <- subset(heart,subset = Broad_celltype == "Neurons")
```

```{r echo=FALSE}
neuron <- readRDS(file="./output/neuron-int-FYA-filtered.Rds")
```


# Check for bad quality cells

Check for cells with very low number of uniquely detected genes. 

```{r}
par(mfrow=c(1,2))
plot(density(neuron$nFeature_RNA),main="Number of genes detected")
abline(v=500,col=2)
plot(density(neuron$nCount_RNA),main="Library size")
abline(v=2500,col=2)

#neuron <- subset(neuron, subset = nFeature_RNA > 500 & nCount_RNA > 2500)
dim(neuron)
table(neuron$biorep)
```


# Run new integration with SCtransform normalisation

There are very few cells for each biological replicate, so I will normalise and integrate the data by group rather than biological replicate.

```{r eval=FALSE}
neuron.list <- SplitObject(neuron, split.by = "orig.ident")
```

```{r eval=FALSE}
for (i in 1:length(neuron.list)) {
    neuron.list[[i]] <- SCTransform(neuron.list[[i]], verbose = FALSE)
}
```

```{r eval=FALSE}
kf <- min(sapply(neuron.list, ncol))
neuron.anchors <- FindIntegrationAnchors(object.list = neuron.list, dims=1:30,anchor.features = 3000,k.filter=kf)
neuron.integrated <- IntegrateData(anchorset = neuron.anchors,dims=1:30)
```

```{r echo=FALSE}
neuron.integrated <- neuron
```


# Perform clustering

```{r}
DefaultAssay(object = neuron.integrated) <- "integrated"
```

## Perform scaling and PCA

```{r eval=FALSE}
neuron.integrated <- ScaleData(neuron.integrated, verbose = FALSE)
neuron.integrated <- RunPCA(neuron.integrated, npcs = 50, verbose = FALSE)
```

```{r}
ElbowPlot(neuron.integrated,ndims=50)
```

```{r,fig.height=10, fig.width=12}
VizDimLoadings(neuron.integrated, dims = 1:4, reduction = "pca")
```

```{r}
DimPlot(neuron.integrated, reduction = "pca",group.by="orig.ident")
DimPlot(neuron.integrated, reduction = "pca",group.by="biorep")
DimPlot(neuron.integrated, reduction = "pca",group.by="sex")
DimPlot(neuron.integrated, reduction = "pca",group.by="batch")
```

```{r,fig.height=10, fig.width=12}
DimHeatmap(neuron.integrated, dims = 1:15, cells = 500, balanced = TRUE)
DimHeatmap(neuron.integrated, dims = 16:30, cells = 500, balanced = TRUE)
#DimHeatmap(neuron.integrated, dims = 31:45, cells = 500, balanced = TRUE)
```

## Perform nearest neighbours clustering

```{r eval=FALSE}
neuron.integrated <- FindNeighbors(neuron.integrated, dims = 1:10)
neuron.integrated <- FindClusters(neuron.integrated, resolution = 0.1)
```

```{r}
table(Idents(neuron.integrated))
```

```{r}
par(mfrow=c(1,1))
par(mar=c(5,4,2,2))
barplot(table(Idents(neuron.integrated)),ylab="Number of cells",xlab="Clusters")
title("Number of cells in each cluster")
```


## Visualisation with TSNE

```{r eval=FALSE}
set.seed(10)
neuron.integrated <- RunTSNE(neuron.integrated, reduction = "pca", dims = 1:10)
```

```{r}
DimPlot(neuron.integrated, reduction = "tsne",label=TRUE,label.size = 6,pt.size = 0.5)+NoLegend()
```

```{r eval=FALSE}
pdf(file="./output/Figures/tsne-neuronALL-res01.pdf",width=10,height=8,onefile = FALSE)
DimPlot(neuron.integrated, reduction = "tsne",label=TRUE,label.size = 6,pt.size = 0.5)+NoLegend()
dev.off()
```

```{r}
DimPlot(neuron.integrated, reduction = "tsne", group.by = "orig.ident")
```

```{r,fig.width=12,fig.height=6}
DimPlot(neuron.integrated, reduction = "tsne", split.by = "orig.ident")
```

```{r}
DimPlot(neuron.integrated, reduction = "tsne", group.by = "biorep")
```


```{r}
DimPlot(neuron.integrated, reduction = "tsne", group.by = "sex")
```

```{r,fig.width=10,fig.height=7}
DimPlot(neuron.integrated, reduction = "tsne", split.by = "sex")
```

```{r}
DimPlot(neuron.integrated, reduction = "tsne", group.by = "batch")
```


```{r,fig.width=12,fig.height=6}
par(mfrow=c(1,1))
par(mar=c(4,4,2,2))
tab <- table(Idents(neuron.integrated),neuron.integrated$biorep)
barplot(t(tab/rowSums(tab)),beside=TRUE,col=ggplotColors(9),legend=TRUE)
```

```{r,fig.width=12,fig.height=6}
par(mfrow=c(1,1))
par(mar=c(4,4,2,2))
tab <- table(Idents(neuron.integrated),neuron.integrated$orig.ident)
barplot(t(tab/rowSums(tab)),beside=TRUE,col=ggplotColors(3))
legend("topleft",legend=colnames(tab),fill=ggplotColors(3))
```

# Visualisation with clustree

```{r eval=FALSE}
clusres <- c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2)
```
```{r eval=FALSE}
for(i in 1:length(clusres)){
  neuron.integrated <- FindClusters(neuron.integrated, 
                                   resolution = clusres[i])
}
```

```{r}
pct.male <- function(x) {mean(x=="m")}
pct.female <- function(x) {mean(x=="f")}
pct.fetal <- function(x) {mean(x=="fetal")}
pct.young <- function(x) {mean(x=="young")}
pct.adult <- function(x) {mean(x=="adult")}
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.")
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.",
         node_colour = "sex", node_colour_aggr = "pct.female",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.",
         node_colour = "sex", node_colour_aggr = "pct.male",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.fetal",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.young",assay="RNA")
```

```{r,fig.width=10,fig.height=10}
clustree(neuron.integrated, prefix = "integrated_snn_res.",
         node_colour = "orig.ident", node_colour_aggr = "pct.adult",assay="RNA")
```


# Save Seurat object

```{r}
DefaultAssay(neuron.integrated) <- "RNA"
Idents(neuron.integrated) <- neuron.integrated$integrated_snn_res.0.1
```

```{r eval=FALSE}
saveRDS(neuron.integrated,file="./output/RDataObjects/neuron-int-FYA-filtered.Rds")
#neuron.integrated <- readRDS(file="./output/RDataObjects/neuron-int-FYA.Rds")
```

```{r eval=FALSE}
# Load unfiltered counts matrix for every sample (object all)
load("./output/RDataObjects/all-counts.Rdata")
```
